3FastThreeFurious
This is a top down car dodging game. The player is the red car and has 3
lives. Every time the player collides with a (random colored car based
on RNG) car he/she will lose a life. At game over, the game will notify
how many points the player achieved. The player achieves 10 points every
time he or she dodges a car and it leaves the screen.

This project was coded in Processing.